requirements.txt
Django==5.2.8
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
drf-spectacular>=0.27.0
python-decouple>=3.8
